test: Compare integers by same signs
`BOOST_*` allows for comparison of integers with different signs. This
is problematic in the case of overflow and underflow, and should not be
valid for a unit test framework.

+++ src/test/sock_tests.cpp
@@ -112,7 +112,7 @@ struct TcpSocketPair {
         sockaddr_in bound{};
         socklen_t blen = sizeof(bound);
         BOOST_REQUIRE_EQUAL(receiver.GetSockName(reinterpret_cast<sockaddr*>(&bound), &blen), 0);
-        BOOST_REQUIRE_EQUAL(blen, sizeof(bound));
+        BOOST_REQUIRE_EQUAL(static_cast<size_t>(blen), sizeof(bound));
 
         BOOST_REQUIRE_EQUAL(sender.Connect(reinterpret_cast<sockaddr*>(&bound), sizeof(bound)), 0);
